Implement two operand instructions
This implements two operand instructions and also performs a significant
amount of refactoring. The API is cleaned up with more clear naming,
code is broken out into functions to make it more testable, and rather
than manually defining all of the individual instructions switch to
using macros to define them.