Merge Source and Destination to support br
The br emulated instruction uses the source operand, rather than the
destination operand, as the destination of the instruction. This is the
only instruction that does this but creates type issues. Instead of
maintaining two separate types, where Destination is a subset of Source,
just let them share the same type and rely on the decoding logic to
ensure that only valid variants are used for a destination.