Provide instruction len for emulated instructions
Deriving the emulated instruction is a lossy process that can lose the
specific original machine code used to implement it due to functionally
equivalent implementations of the same instruction. To ensure that
downstream consumers can keep track of where the cursor is in the code
segment we need to provide the correct value.