Store original instruction for emulated
Rather than store just the length of the emulated instruction so that
our offsets are correct, let's store the actual instruction so that we
can ensure that we're able to re-create the actual instruction when we
support assembling. With only the length we wont be able to re-create
with certainty.