Change Operand::Immediate to u16
Change from an i16 to a u16. Immediates need to be able to hold a full
16 bit unsigned value which an i16 can't. For display purposes we may
need to do some heuristics to correctly determine what the correct value
to display is but this led to bad casts when working with real firmware
images and attempting to handle the value correctly. This is likely the
correct way to represent the data.

Closes #8